requirements.txt
transformers
fastapi
uvicorn
python-dotenv